Force the sphinx extension to rebuild specs.
When Sphinx loads the sources, it swallows errors than can occur
when building a spec. This can lead to having an incomplete spec
in the documentation. To work around this, we delete the '_spec'
attribute in the sphinx extension, which means the spec will have
to be re-built at that point. Errors will then be raised, which makes
them easier to find and debug.